requirements.txt
transformers
keybert
pandas
sqlalchemy
nltk
spacy
torch
protobuf
seaborn
mplcursors
scikit-learn
sentencepiece
wordcloud
langdetect
accelerate
matplotlib
pyperclip
faiss-cpu
jq
ipython
jupyter
ipykernel
setuptools
wheel
python-Levenshtein
easy_llama